# Optimize component part construction