Component part construction performs some redundant conversions on constants.

The conversions are more costly than they look, as the decimal to integer path range checks against type ranges on the integer types, which in turn performs conversions of these to the decimal type. For `Dec128/UInt128` this is particularly bad as the integer to decimal conversion falls back to conversion via strings.

Fixed by precomputing integer constants.

Simple example benchmark in the form of
```Julia
@btime T($(Ref(123))[], $(Ref(-8))[])
```

| T | before | after |
| --- | --- | --- |
| Dec32 | 10.701 ns (0 allocations: 0 bytes) | 4.200 ns (0 allocations: 0 bytes) |
| Dec64 | 11.912 ns (0 allocations: 0 bytes)  | 3.810 ns (0 allocations: 0 bytes) |
| Dec128 | 394.069 ns (6 allocations: 288 bytes) | 6.510 ns (0 allocations: 0 bytes) |