@ancapdev 2022-03-25
> Would be good to add tests of the form:
> 
> ```julia
> @test DecFP._int_maxintfloat(Dec32) === UInt32(maxintfloat(Dec32))
> @test DecFP._int_maxintfloat(Dec64) === UInt64(maxintfloat(Dec64))
> @test DecFP._int_maxintfloat(Dec128) === UInt128(maxintfloat(Dec128))
> ```

Makes sense, done